requirements.txt
flask==3.0.2
flask-cors==4.0.0
gspread==6.0.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1 